try to fix example

+++ examples/short_tests.jl
@@ -107,7 +107,8 @@
         colormap = [(:black, 0.2), :red, :blue, :green, (:black, 0.2)],
         colorrange = (0.2, 0.8)
     )
-    @cell lines(Circle(Point2f0(0), 1f0); scale_plot = false, resolution = (800, 1000)) # currently not working, fix is in PR
+
+    @cell lines(Circle(Point2f0(0), Float32(1)); scale_plot = false, resolution = (800, 1000))
 
     @cell begin
         v(x::Point2{T}) where T = Point2{T}(x[2], 4*x[1])
